feat(#133): fix all the code offences
